requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
langchain==0.0.350
chromadb==0.4.18
sentence-transformers==2.2.2
pandas==2.1.3
numpy==1.26.2 